requirements.txt
supabase
python-dotenv
pandas
pyarrow
matplotlib
seaborn
scikit-learn
nfl_data_py
cfbd
thefuzz
python-Levenshtein
jupyter
